requirements.txt
numpy==1.24.3
pandas==2.0.3
Pillow==11.3.0
roifile==2023.5.12
xlsxwriter==3.2.5
